TwitterApp
This project has been developed using JavaScript and HTML. Originally, this application was designed as a user data retrieval and filtering application using the Twitter API. Due to issues encountered during the development phase with the Twitter API usage, user data was retrieved from a JSON file, and various manipulations and filters were applied to the data.

User and Twitter filters are available within the project. You can use this project as an example for your school lessons and in projects you develop.